Split generated suites into files based on suite name
