--- spark/src/test/scala/org/apache/spark/sql/delta/generatedsuites/DeleteSuites1.scala
@@ -1,234 +1,0 @@
-/*
- * Copyright (2021) The Delta Lake Project Authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-// ***********************************************************************************
-// * This file is automatically generated. Manual modification is not allowed.       *
-// * There is a unit test that should prevent merging a manual change.               *
-// *                                                                                 *
-// * To make changes to the suites, modify the generator script config at            *
-// * SuiteGeneratorConfig.scala and run it. The generator can be run via the         *
-// * sbt command deltaSuiteGenerator / run.                                          *
-// *                                                                                 *
-// * DO NOT TOUCH ANYTHING IN THIS FILE!                                             *
-// ***********************************************************************************
-
-// scalastyle:off line.size.limit
-package org.apache.spark.sql.delta.generatedsuites
-
-import org.apache.spark.sql.delta._
-import org.apache.spark.sql.delta.cdc._
-import org.apache.spark.sql.delta.rowid._
-
-class DeleteScalaScalaSuite extends DeleteScalaTests with DeleteScalaMixin
-class DeleteBaseScalaSuite extends DeleteBaseTests with DeleteScalaMixin
-
-class DeleteCDCSQLPathBasedCDCOnSuite
-  extends DeleteCDCTests
-  with DeleteSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with DeleteCDCMixin
-
-class DeleteSQLSQLNameBasedSuite
-  extends DeleteSQLTests
-  with DeleteSQLMixin
-  with DeltaDMLTestUtilsNameBased
-
-class DeleteSQLSQLPathBasedSuite
-  extends DeleteSQLTests
-  with DeleteSQLMixin
-  with DeltaDMLTestUtilsPathBased
-
-class DeleteSQLSQLPathBasedColMapIdModeSuite
-  extends DeleteSQLTests
-  with DeleteSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with DeltaColumnMappingEnableIdMode
-
-class DeleteSQLSQLPathBasedColMapNameModeSuite
-  extends DeleteSQLTests
-  with DeleteSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with DeltaColumnMappingEnableNameMode
-  with DeleteSQLNameColumnMappingMixin
-
-class DeleteSQLSQLPathBasedDVPredPushOffSuite
-  extends DeleteSQLTests
-  with DeleteSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with DeleteSQLWithDeletionVectorsMixin
-  with PredicatePushdownDisabled
-
-class DeleteSQLSQLPathBasedDVPredPushOnSuite
-  extends DeleteSQLTests
-  with DeleteSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with DeleteSQLWithDeletionVectorsMixin
-  with PredicatePushdownEnabled
-
-class DeleteSQLSQLPathBasedCDCOnSuite
-  extends DeleteSQLTests
-  with DeleteSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with DeleteCDCMixin
-
-class DeleteTempViewSQLNameBasedSuite
-  extends DeleteTempViewTests
-  with DeleteSQLMixin
-  with DeltaDMLTestUtilsNameBased
-
-class DeleteTempViewSQLPathBasedSuite
-  extends DeleteTempViewTests
-  with DeleteSQLMixin
-  with DeltaDMLTestUtilsPathBased
-
-class DeleteTempViewSQLPathBasedColMapIdModeSuite
-  extends DeleteTempViewTests
-  with DeleteSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with DeltaColumnMappingEnableIdMode
-
-class DeleteTempViewSQLPathBasedColMapNameModeSuite
-  extends DeleteTempViewTests
-  with DeleteSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with DeltaColumnMappingEnableNameMode
-  with DeleteSQLNameColumnMappingMixin
-
-class DeleteTempViewSQLPathBasedDVPredPushOffSuite
-  extends DeleteTempViewTests
-  with DeleteSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with DeleteSQLWithDeletionVectorsMixin
-  with PredicatePushdownDisabled
-
-class DeleteTempViewSQLPathBasedDVPredPushOnSuite
-  extends DeleteTempViewTests
-  with DeleteSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with DeleteSQLWithDeletionVectorsMixin
-  with PredicatePushdownEnabled
-
-class DeleteTempViewSQLPathBasedCDCOnSuite
-  extends DeleteTempViewTests
-  with DeleteSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with DeleteCDCMixin
-
-class DeleteBaseSQLNameBasedSuite
-  extends DeleteBaseTests
-  with DeleteSQLMixin
-  with DeltaDMLTestUtilsNameBased
-
-class DeleteBaseSQLPathBasedSuite
-  extends DeleteBaseTests
-  with DeleteSQLMixin
-  with DeltaDMLTestUtilsPathBased
-
-class DeleteBaseSQLPathBasedColMapIdModeSuite
-  extends DeleteBaseTests
-  with DeleteSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with DeltaColumnMappingEnableIdMode
-
-class DeleteBaseSQLPathBasedColMapNameModeSuite
-  extends DeleteBaseTests
-  with DeleteSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with DeltaColumnMappingEnableNameMode
-  with DeleteSQLNameColumnMappingMixin
-
-class DeleteBaseSQLPathBasedDVPredPushOffSuite
-  extends DeleteBaseTests
-  with DeleteSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with DeleteSQLWithDeletionVectorsMixin
-  with PredicatePushdownDisabled
-
-class DeleteBaseSQLPathBasedDVPredPushOnSuite
-  extends DeleteBaseTests
-  with DeleteSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with DeleteSQLWithDeletionVectorsMixin
-  with PredicatePushdownEnabled
-
-class DeleteBaseSQLPathBasedCDCOnSuite
-  extends DeleteBaseTests
-  with DeleteSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with DeleteCDCMixin
-
-class RowTrackingDeleteSuiteBaseDVOffSuite
-  extends RowTrackingDeleteSuiteBase
-  with PersistentDVDisabled
-
-class RowTrackingDeleteSuiteBaseDVOnSuite
-  extends RowTrackingDeleteSuiteBase
-  with PersistentDVEnabled
-
-class RowTrackingDeleteSuiteBaseCDCOnDVOffSuite
-  extends RowTrackingDeleteSuiteBase
-  with CDCEnabled
-  with PersistentDVDisabled
-
-class RowTrackingDeleteSuiteBaseCDCOnDVOnSuite
-  extends RowTrackingDeleteSuiteBase
-  with CDCEnabled
-  with PersistentDVEnabled
-
-class RowTrackingDeleteSuiteBaseDVOffColMapIdModeSuite
-  extends RowTrackingDeleteSuiteBase
-  with PersistentDVDisabled
-  with DeltaColumnMappingEnableIdMode
-
-class RowTrackingDeleteSuiteBaseDVOffColMapNameModeSuite
-  extends RowTrackingDeleteSuiteBase
-  with PersistentDVDisabled
-  with DeltaColumnMappingEnableNameMode
-
-class RowTrackingDeleteSuiteBaseCDCOnDVOnColMapIdModeSuite
-  extends RowTrackingDeleteSuiteBase
-  with CDCEnabled
-  with PersistentDVEnabled
-  with DeltaColumnMappingEnableIdMode
-
-class RowTrackingDeleteSuiteBaseCDCOnDVOnColMapNameModeSuite
-  extends RowTrackingDeleteSuiteBase
-  with CDCEnabled
-  with PersistentDVEnabled
-  with DeltaColumnMappingEnableNameMode
-
-class RowTrackingDeleteDvBaseDVOnSuite extends RowTrackingDeleteDvBase with PersistentDVEnabled
-
-class RowTrackingDeleteDvBaseCDCOnDVOnSuite
-  extends RowTrackingDeleteDvBase
-  with CDCEnabled
-  with PersistentDVEnabled
-
-class RowTrackingDeleteDvBaseCDCOnDVOnColMapIdModeSuite
-  extends RowTrackingDeleteDvBase
-  with CDCEnabled
-  with PersistentDVEnabled
-  with DeltaColumnMappingEnableIdMode
-
-class RowTrackingDeleteDvBaseCDCOnDVOnColMapNameModeSuite
-  extends RowTrackingDeleteDvBase
-  with CDCEnabled
-  with PersistentDVEnabled
-  with DeltaColumnMappingEnableNameMode
