--- spark/src/test/scala/org/apache/spark/sql/delta/generatedsuites/MergeSuites1.scala
@@ -1,699 +1,0 @@
-/*
- * Copyright (2021) The Delta Lake Project Authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-// ***********************************************************************************
-// * This file is automatically generated. Manual modification is not allowed.       *
-// * There is a unit test that should prevent merging a manual change.               *
-// *                                                                                 *
-// * To make changes to the suites, modify the generator script config at            *
-// * SuiteGeneratorConfig.scala and run it. The generator can be run via the         *
-// * sbt command deltaSuiteGenerator / run.                                          *
-// *                                                                                 *
-// * DO NOT TOUCH ANYTHING IN THIS FILE!                                             *
-// ***********************************************************************************
-
-// scalastyle:off line.size.limit
-package org.apache.spark.sql.delta.generatedsuites
-
-import org.apache.spark.sql.delta._
-import org.apache.spark.sql.delta.cdc._
-import org.apache.spark.sql.delta.rowid._
-
-class MergeIntoScalaScalaSuite extends MergeIntoScalaTests with MergeIntoScalaMixin
-class MergeIntoBasicScalaSuite extends MergeIntoBasicTests with MergeIntoScalaMixin
-class MergeIntoNestedDataScalaSuite extends MergeIntoNestedDataTests with MergeIntoScalaMixin
-
-class MergeIntoUnlimitedMergeClausesScalaSuite
-  extends MergeIntoUnlimitedMergeClausesTests
-  with MergeIntoScalaMixin
-
-class MergeIntoAnalysisExceptionScalaSuite
-  extends MergeIntoAnalysisExceptionTests
-  with MergeIntoScalaMixin
-
-class MergeIntoExtendedSyntaxScalaSuite
-  extends MergeIntoExtendedSyntaxTests
-  with MergeIntoScalaMixin
-
-class MergeIntoSuiteBaseMiscScalaSuite extends MergeIntoSuiteBaseMiscTests with MergeIntoScalaMixin
-
-class MergeIntoNotMatchedBySourceScalaSuite
-  extends MergeIntoNotMatchedBySourceSuite
-  with MergeIntoScalaMixin
-
-class MergeIntoNotMatchedBySourceCDCPart1ScalaSuite
-  extends MergeIntoNotMatchedBySourceCDCPart1Tests
-  with MergeIntoScalaMixin
-
-class MergeIntoNotMatchedBySourceCDCPart2ScalaSuite
-  extends MergeIntoNotMatchedBySourceCDCPart2Tests
-  with MergeIntoScalaMixin
-
-class MergeIntoSchemaEvolutionCoreScalaSuite
-  extends MergeIntoSchemaEvolutionCoreTests
-  with MergeIntoScalaMixin
-
-class MergeIntoSchemaEvolutionBaseScalaSuite
-  extends MergeIntoSchemaEvolutionBaseTests
-  with MergeIntoScalaMixin
-
-class MergeIntoSchemaEvolutionStoreAssignmentPolicyScalaSuite
-  extends MergeIntoSchemaEvolutionStoreAssignmentPolicyTests
-  with MergeIntoScalaMixin
-
-class MergeIntoSchemaEvolutionNotMatchedBySourceScalaSuite
-  extends MergeIntoSchemaEvolutionNotMatchedBySourceTests
-  with MergeIntoScalaMixin
-
-class MergeIntoNestedStructInMapEvolutionScalaSuite
-  extends MergeIntoNestedStructInMapEvolutionTests
-  with MergeIntoScalaMixin
-
-class MergeIntoNestedStructEvolutionScalaSuite
-  extends MergeIntoNestedStructEvolutionTests
-  with MergeIntoScalaMixin
-
-class MergeCDCSQLPathBasedCDCOnSuite
-  extends MergeCDCTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeCDCMixin
-
-class MergeCDCSQLPathBasedCDCOnDVsPredPushOffSuite
-  extends MergeCDCTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeIntoDVsMixin
-  with PredicatePushdownDisabled
-  with MergeCDCMixin
-  with MergeCDCWithDVsMixin
-
-class MergeCDCSQLPathBasedCDCOnDVsPredPushOnSuite
-  extends MergeCDCTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeIntoDVsMixin
-  with PredicatePushdownEnabled
-  with MergeCDCMixin
-  with MergeCDCWithDVsMixin
-
-class MergeIntoDVsSQLPathBasedDVsPredPushOffSuite
-  extends MergeIntoDVsTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with MergeIntoDVsMixin
-  with PredicatePushdownDisabled
-
-class MergeIntoDVsSQLPathBasedDVsPredPushOnSuite
-  extends MergeIntoDVsTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with MergeIntoDVsMixin
-  with PredicatePushdownEnabled
-
-class MergeIntoDVsSQLPathBasedCDCOnDVsPredPushOffSuite
-  extends MergeIntoDVsTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeIntoDVsMixin
-  with PredicatePushdownDisabled
-  with MergeCDCMixin
-  with MergeCDCWithDVsMixin
-
-class MergeIntoDVsSQLPathBasedCDCOnDVsPredPushOnSuite
-  extends MergeIntoDVsTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeIntoDVsMixin
-  with PredicatePushdownEnabled
-  with MergeCDCMixin
-  with MergeCDCWithDVsMixin
-
-class MergeIntoSQLSQLNameBasedSuite
-  extends MergeIntoSQLTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsNameBased
-
-class MergeIntoSQLSQLPathBasedSuite
-  extends MergeIntoSQLTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-
-class MergeIntoSQLSQLPathBasedColMapIdModeSuite
-  extends MergeIntoSQLTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with DeltaColumnMappingEnableIdMode
-  with MergeIntoSQLColumnMappingOverrides
-
-class MergeIntoSQLSQLPathBasedColMapNameModeSuite
-  extends MergeIntoSQLTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with DeltaColumnMappingEnableNameMode
-  with MergeIntoSQLColumnMappingOverrides
-
-class MergeIntoSQLSQLPathBasedDVsPredPushOffSuite
-  extends MergeIntoSQLTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with MergeIntoDVsMixin
-  with PredicatePushdownDisabled
-
-class MergeIntoSQLSQLPathBasedDVsPredPushOnSuite
-  extends MergeIntoSQLTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with MergeIntoDVsMixin
-  with PredicatePushdownEnabled
-
-class MergeIntoSQLSQLPathBasedCDCOnSuite
-  extends MergeIntoSQLTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeCDCMixin
-
-class MergeIntoSQLSQLPathBasedCDCOnDVsPredPushOffSuite
-  extends MergeIntoSQLTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeIntoDVsMixin
-  with PredicatePushdownDisabled
-  with MergeCDCMixin
-  with MergeCDCWithDVsMixin
-
-class MergeIntoSQLSQLPathBasedCDCOnDVsPredPushOnSuite
-  extends MergeIntoSQLTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeIntoDVsMixin
-  with PredicatePushdownEnabled
-  with MergeCDCMixin
-  with MergeCDCWithDVsMixin
-
-class MergeIntoSQLNondeterministicOrderSQLNameBasedSuite
-  extends MergeIntoSQLNondeterministicOrderTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsNameBased
-
-class MergeIntoSQLNondeterministicOrderSQLPathBasedSuite
-  extends MergeIntoSQLNondeterministicOrderTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-
-class MergeIntoSQLNondeterministicOrderSQLPathBasedColMapIdModeSuite
-  extends MergeIntoSQLNondeterministicOrderTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with DeltaColumnMappingEnableIdMode
-  with MergeIntoSQLColumnMappingOverrides
-
-class MergeIntoSQLNondeterministicOrderSQLPathBasedColMapNameModeSuite
-  extends MergeIntoSQLNondeterministicOrderTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with DeltaColumnMappingEnableNameMode
-  with MergeIntoSQLColumnMappingOverrides
-
-class MergeIntoSQLNondeterministicOrderSQLPathBasedDVsPredPushOffSuite
-  extends MergeIntoSQLNondeterministicOrderTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with MergeIntoDVsMixin
-  with PredicatePushdownDisabled
-
-class MergeIntoSQLNondeterministicOrderSQLPathBasedDVsPredPushOnSuite
-  extends MergeIntoSQLNondeterministicOrderTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with MergeIntoDVsMixin
-  with PredicatePushdownEnabled
-
-class MergeIntoSQLNondeterministicOrderSQLPathBasedCDCOnSuite
-  extends MergeIntoSQLNondeterministicOrderTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeCDCMixin
-
-class MergeIntoSQLNondeterministicOrderSQLPathBasedCDCOnDVsPredPushOffSuite
-  extends MergeIntoSQLNondeterministicOrderTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeIntoDVsMixin
-  with PredicatePushdownDisabled
-  with MergeCDCMixin
-  with MergeCDCWithDVsMixin
-
-class MergeIntoSQLNondeterministicOrderSQLPathBasedCDCOnDVsPredPushOnSuite
-  extends MergeIntoSQLNondeterministicOrderTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeIntoDVsMixin
-  with PredicatePushdownEnabled
-  with MergeCDCMixin
-  with MergeCDCWithDVsMixin
-
-class MergeIntoBasicSQLNameBasedSuite
-  extends MergeIntoBasicTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsNameBased
-
-class MergeIntoBasicSQLPathBasedSuite
-  extends MergeIntoBasicTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-
-class MergeIntoBasicSQLPathBasedColMapIdModeSuite
-  extends MergeIntoBasicTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with DeltaColumnMappingEnableIdMode
-  with MergeIntoSQLColumnMappingOverrides
-
-class MergeIntoBasicSQLPathBasedColMapNameModeSuite
-  extends MergeIntoBasicTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with DeltaColumnMappingEnableNameMode
-  with MergeIntoSQLColumnMappingOverrides
-
-class MergeIntoBasicSQLPathBasedDVsPredPushOffSuite
-  extends MergeIntoBasicTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with MergeIntoDVsMixin
-  with PredicatePushdownDisabled
-
-class MergeIntoBasicSQLPathBasedDVsPredPushOnSuite
-  extends MergeIntoBasicTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with MergeIntoDVsMixin
-  with PredicatePushdownEnabled
-
-class MergeIntoBasicSQLPathBasedCDCOnSuite
-  extends MergeIntoBasicTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeCDCMixin
-
-class MergeIntoBasicSQLPathBasedCDCOnDVsPredPushOffSuite
-  extends MergeIntoBasicTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeIntoDVsMixin
-  with PredicatePushdownDisabled
-  with MergeCDCMixin
-  with MergeCDCWithDVsMixin
-
-class MergeIntoBasicSQLPathBasedCDCOnDVsPredPushOnSuite
-  extends MergeIntoBasicTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeIntoDVsMixin
-  with PredicatePushdownEnabled
-  with MergeCDCMixin
-  with MergeCDCWithDVsMixin
-
-class MergeIntoTempViewsSQLNameBasedSuite
-  extends MergeIntoTempViewsTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsNameBased
-
-class MergeIntoTempViewsSQLPathBasedSuite
-  extends MergeIntoTempViewsTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-
-class MergeIntoTempViewsSQLPathBasedColMapIdModeSuite
-  extends MergeIntoTempViewsTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with DeltaColumnMappingEnableIdMode
-  with MergeIntoSQLColumnMappingOverrides
-
-class MergeIntoTempViewsSQLPathBasedColMapNameModeSuite
-  extends MergeIntoTempViewsTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with DeltaColumnMappingEnableNameMode
-  with MergeIntoSQLColumnMappingOverrides
-
-class MergeIntoTempViewsSQLPathBasedDVsPredPushOffSuite
-  extends MergeIntoTempViewsTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with MergeIntoDVsMixin
-  with PredicatePushdownDisabled
-
-class MergeIntoTempViewsSQLPathBasedDVsPredPushOnSuite
-  extends MergeIntoTempViewsTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with MergeIntoDVsMixin
-  with PredicatePushdownEnabled
-
-class MergeIntoTempViewsSQLPathBasedCDCOnSuite
-  extends MergeIntoTempViewsTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeCDCMixin
-
-class MergeIntoTempViewsSQLPathBasedCDCOnDVsPredPushOffSuite
-  extends MergeIntoTempViewsTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeIntoDVsMixin
-  with PredicatePushdownDisabled
-  with MergeCDCMixin
-  with MergeCDCWithDVsMixin
-
-class MergeIntoTempViewsSQLPathBasedCDCOnDVsPredPushOnSuite
-  extends MergeIntoTempViewsTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeIntoDVsMixin
-  with PredicatePushdownEnabled
-  with MergeCDCMixin
-  with MergeCDCWithDVsMixin
-
-class MergeIntoNestedDataSQLNameBasedSuite
-  extends MergeIntoNestedDataTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsNameBased
-
-class MergeIntoNestedDataSQLPathBasedSuite
-  extends MergeIntoNestedDataTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-
-class MergeIntoNestedDataSQLPathBasedColMapIdModeSuite
-  extends MergeIntoNestedDataTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with DeltaColumnMappingEnableIdMode
-  with MergeIntoSQLColumnMappingOverrides
-
-class MergeIntoNestedDataSQLPathBasedColMapNameModeSuite
-  extends MergeIntoNestedDataTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with DeltaColumnMappingEnableNameMode
-  with MergeIntoSQLColumnMappingOverrides
-
-class MergeIntoNestedDataSQLPathBasedDVsPredPushOffSuite
-  extends MergeIntoNestedDataTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with MergeIntoDVsMixin
-  with PredicatePushdownDisabled
-
-class MergeIntoNestedDataSQLPathBasedDVsPredPushOnSuite
-  extends MergeIntoNestedDataTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with MergeIntoDVsMixin
-  with PredicatePushdownEnabled
-
-class MergeIntoNestedDataSQLPathBasedCDCOnSuite
-  extends MergeIntoNestedDataTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeCDCMixin
-
-class MergeIntoNestedDataSQLPathBasedCDCOnDVsPredPushOffSuite
-  extends MergeIntoNestedDataTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeIntoDVsMixin
-  with PredicatePushdownDisabled
-  with MergeCDCMixin
-  with MergeCDCWithDVsMixin
-
-class MergeIntoNestedDataSQLPathBasedCDCOnDVsPredPushOnSuite
-  extends MergeIntoNestedDataTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeIntoDVsMixin
-  with PredicatePushdownEnabled
-  with MergeCDCMixin
-  with MergeCDCWithDVsMixin
-
-class MergeIntoUnlimitedMergeClausesSQLNameBasedSuite
-  extends MergeIntoUnlimitedMergeClausesTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsNameBased
-
-class MergeIntoUnlimitedMergeClausesSQLPathBasedSuite
-  extends MergeIntoUnlimitedMergeClausesTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-
-class MergeIntoUnlimitedMergeClausesSQLPathBasedColMapIdModeSuite
-  extends MergeIntoUnlimitedMergeClausesTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with DeltaColumnMappingEnableIdMode
-  with MergeIntoSQLColumnMappingOverrides
-
-class MergeIntoUnlimitedMergeClausesSQLPathBasedColMapNameModeSuite
-  extends MergeIntoUnlimitedMergeClausesTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with DeltaColumnMappingEnableNameMode
-  with MergeIntoSQLColumnMappingOverrides
-
-class MergeIntoUnlimitedMergeClausesSQLPathBasedDVsPredPushOffSuite
-  extends MergeIntoUnlimitedMergeClausesTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with MergeIntoDVsMixin
-  with PredicatePushdownDisabled
-
-class MergeIntoUnlimitedMergeClausesSQLPathBasedDVsPredPushOnSuite
-  extends MergeIntoUnlimitedMergeClausesTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with MergeIntoDVsMixin
-  with PredicatePushdownEnabled
-
-class MergeIntoUnlimitedMergeClausesSQLPathBasedCDCOnSuite
-  extends MergeIntoUnlimitedMergeClausesTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeCDCMixin
-
-class MergeIntoUnlimitedMergeClausesSQLPathBasedCDCOnDVsPredPushOffSuite
-  extends MergeIntoUnlimitedMergeClausesTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeIntoDVsMixin
-  with PredicatePushdownDisabled
-  with MergeCDCMixin
-  with MergeCDCWithDVsMixin
-
-class MergeIntoUnlimitedMergeClausesSQLPathBasedCDCOnDVsPredPushOnSuite
-  extends MergeIntoUnlimitedMergeClausesTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeIntoDVsMixin
-  with PredicatePushdownEnabled
-  with MergeCDCMixin
-  with MergeCDCWithDVsMixin
-
-class MergeIntoAnalysisExceptionSQLNameBasedSuite
-  extends MergeIntoAnalysisExceptionTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsNameBased
-
-class MergeIntoAnalysisExceptionSQLPathBasedSuite
-  extends MergeIntoAnalysisExceptionTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-
-class MergeIntoAnalysisExceptionSQLPathBasedColMapIdModeSuite
-  extends MergeIntoAnalysisExceptionTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with DeltaColumnMappingEnableIdMode
-  with MergeIntoSQLColumnMappingOverrides
-
-class MergeIntoAnalysisExceptionSQLPathBasedColMapNameModeSuite
-  extends MergeIntoAnalysisExceptionTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with DeltaColumnMappingEnableNameMode
-  with MergeIntoSQLColumnMappingOverrides
-
-class MergeIntoAnalysisExceptionSQLPathBasedDVsPredPushOffSuite
-  extends MergeIntoAnalysisExceptionTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with MergeIntoDVsMixin
-  with PredicatePushdownDisabled
-
-class MergeIntoAnalysisExceptionSQLPathBasedDVsPredPushOnSuite
-  extends MergeIntoAnalysisExceptionTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with MergeIntoDVsMixin
-  with PredicatePushdownEnabled
-
-class MergeIntoAnalysisExceptionSQLPathBasedCDCOnSuite
-  extends MergeIntoAnalysisExceptionTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeCDCMixin
-
-class MergeIntoAnalysisExceptionSQLPathBasedCDCOnDVsPredPushOffSuite
-  extends MergeIntoAnalysisExceptionTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeIntoDVsMixin
-  with PredicatePushdownDisabled
-  with MergeCDCMixin
-  with MergeCDCWithDVsMixin
-
-class MergeIntoAnalysisExceptionSQLPathBasedCDCOnDVsPredPushOnSuite
-  extends MergeIntoAnalysisExceptionTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeIntoDVsMixin
-  with PredicatePushdownEnabled
-  with MergeCDCMixin
-  with MergeCDCWithDVsMixin
-
-class MergeIntoExtendedSyntaxSQLNameBasedSuite
-  extends MergeIntoExtendedSyntaxTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsNameBased
-
-class MergeIntoExtendedSyntaxSQLPathBasedSuite
-  extends MergeIntoExtendedSyntaxTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-
-class MergeIntoExtendedSyntaxSQLPathBasedColMapIdModeSuite
-  extends MergeIntoExtendedSyntaxTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with DeltaColumnMappingEnableIdMode
-  with MergeIntoSQLColumnMappingOverrides
-
-class MergeIntoExtendedSyntaxSQLPathBasedColMapNameModeSuite
-  extends MergeIntoExtendedSyntaxTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with DeltaColumnMappingEnableNameMode
-  with MergeIntoSQLColumnMappingOverrides
-
-class MergeIntoExtendedSyntaxSQLPathBasedDVsPredPushOffSuite
-  extends MergeIntoExtendedSyntaxTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with MergeIntoDVsMixin
-  with PredicatePushdownDisabled
-
-class MergeIntoExtendedSyntaxSQLPathBasedDVsPredPushOnSuite
-  extends MergeIntoExtendedSyntaxTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with MergeIntoDVsMixin
-  with PredicatePushdownEnabled
-
-class MergeIntoExtendedSyntaxSQLPathBasedCDCOnSuite
-  extends MergeIntoExtendedSyntaxTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeCDCMixin
-
-class MergeIntoExtendedSyntaxSQLPathBasedCDCOnDVsPredPushOffSuite
-  extends MergeIntoExtendedSyntaxTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeIntoDVsMixin
-  with PredicatePushdownDisabled
-  with MergeCDCMixin
-  with MergeCDCWithDVsMixin
-
-class MergeIntoExtendedSyntaxSQLPathBasedCDCOnDVsPredPushOnSuite
-  extends MergeIntoExtendedSyntaxTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeIntoDVsMixin
-  with PredicatePushdownEnabled
-  with MergeCDCMixin
-  with MergeCDCWithDVsMixin
-
-class MergeIntoSuiteBaseMiscSQLNameBasedSuite
-  extends MergeIntoSuiteBaseMiscTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsNameBased
-
-class MergeIntoSuiteBaseMiscSQLPathBasedSuite
-  extends MergeIntoSuiteBaseMiscTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-
-class MergeIntoSuiteBaseMiscSQLPathBasedColMapIdModeSuite
-  extends MergeIntoSuiteBaseMiscTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with DeltaColumnMappingEnableIdMode
-  with MergeIntoSQLColumnMappingOverrides
-
-class MergeIntoSuiteBaseMiscSQLPathBasedColMapNameModeSuite
-  extends MergeIntoSuiteBaseMiscTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with DeltaColumnMappingEnableNameMode
-  with MergeIntoSQLColumnMappingOverrides
-
-class MergeIntoSuiteBaseMiscSQLPathBasedDVsPredPushOffSuite
-  extends MergeIntoSuiteBaseMiscTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with MergeIntoDVsMixin
-  with PredicatePushdownDisabled
