--- spark/src/test/scala/org/apache/spark/sql/delta/generatedsuites/MergeSuites2.scala
@@ -1,717 +1,0 @@
-/*
- * Copyright (2021) The Delta Lake Project Authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-// ***********************************************************************************
-// * This file is automatically generated. Manual modification is not allowed.       *
-// * There is a unit test that should prevent merging a manual change.               *
-// *                                                                                 *
-// * To make changes to the suites, modify the generator script config at            *
-// * SuiteGeneratorConfig.scala and run it. The generator can be run via the         *
-// * sbt command deltaSuiteGenerator / run.                                          *
-// *                                                                                 *
-// * DO NOT TOUCH ANYTHING IN THIS FILE!                                             *
-// ***********************************************************************************
-
-// scalastyle:off line.size.limit
-package org.apache.spark.sql.delta.generatedsuites
-
-import org.apache.spark.sql.delta._
-import org.apache.spark.sql.delta.cdc._
-import org.apache.spark.sql.delta.rowid._
-
-class MergeIntoSuiteBaseMiscSQLPathBasedDVsPredPushOnSuite
-  extends MergeIntoSuiteBaseMiscTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with MergeIntoDVsMixin
-  with PredicatePushdownEnabled
-
-class MergeIntoSuiteBaseMiscSQLPathBasedCDCOnSuite
-  extends MergeIntoSuiteBaseMiscTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeCDCMixin
-
-class MergeIntoSuiteBaseMiscSQLPathBasedCDCOnDVsPredPushOffSuite
-  extends MergeIntoSuiteBaseMiscTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeIntoDVsMixin
-  with PredicatePushdownDisabled
-  with MergeCDCMixin
-  with MergeCDCWithDVsMixin
-
-class MergeIntoSuiteBaseMiscSQLPathBasedCDCOnDVsPredPushOnSuite
-  extends MergeIntoSuiteBaseMiscTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeIntoDVsMixin
-  with PredicatePushdownEnabled
-  with MergeCDCMixin
-  with MergeCDCWithDVsMixin
-
-class MergeIntoNotMatchedBySourceSQLNameBasedSuite
-  extends MergeIntoNotMatchedBySourceSuite
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsNameBased
-
-class MergeIntoNotMatchedBySourceSQLPathBasedSuite
-  extends MergeIntoNotMatchedBySourceSuite
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-
-class MergeIntoNotMatchedBySourceSQLPathBasedColMapIdModeSuite
-  extends MergeIntoNotMatchedBySourceSuite
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with DeltaColumnMappingEnableIdMode
-  with MergeIntoSQLColumnMappingOverrides
-
-class MergeIntoNotMatchedBySourceSQLPathBasedColMapNameModeSuite
-  extends MergeIntoNotMatchedBySourceSuite
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with DeltaColumnMappingEnableNameMode
-  with MergeIntoSQLColumnMappingOverrides
-
-class MergeIntoNotMatchedBySourceSQLPathBasedDVsPredPushOffSuite
-  extends MergeIntoNotMatchedBySourceSuite
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with MergeIntoDVsMixin
-  with PredicatePushdownDisabled
-
-class MergeIntoNotMatchedBySourceSQLPathBasedDVsPredPushOnSuite
-  extends MergeIntoNotMatchedBySourceSuite
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with MergeIntoDVsMixin
-  with PredicatePushdownEnabled
-
-class MergeIntoNotMatchedBySourceSQLPathBasedCDCOnSuite
-  extends MergeIntoNotMatchedBySourceSuite
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeCDCMixin
-
-class MergeIntoNotMatchedBySourceSQLPathBasedCDCOnDVsPredPushOffSuite
-  extends MergeIntoNotMatchedBySourceSuite
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeIntoDVsMixin
-  with PredicatePushdownDisabled
-  with MergeCDCMixin
-  with MergeCDCWithDVsMixin
-
-class MergeIntoNotMatchedBySourceSQLPathBasedCDCOnDVsPredPushOnSuite
-  extends MergeIntoNotMatchedBySourceSuite
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeIntoDVsMixin
-  with PredicatePushdownEnabled
-  with MergeCDCMixin
-  with MergeCDCWithDVsMixin
-
-class MergeIntoNotMatchedBySourceCDCPart1SQLNameBasedSuite
-  extends MergeIntoNotMatchedBySourceCDCPart1Tests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsNameBased
-
-class MergeIntoNotMatchedBySourceCDCPart1SQLPathBasedSuite
-  extends MergeIntoNotMatchedBySourceCDCPart1Tests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-
-class MergeIntoNotMatchedBySourceCDCPart1SQLPathBasedColMapIdModeSuite
-  extends MergeIntoNotMatchedBySourceCDCPart1Tests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with DeltaColumnMappingEnableIdMode
-  with MergeIntoSQLColumnMappingOverrides
-
-class MergeIntoNotMatchedBySourceCDCPart1SQLPathBasedColMapNameModeSuite
-  extends MergeIntoNotMatchedBySourceCDCPart1Tests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with DeltaColumnMappingEnableNameMode
-  with MergeIntoSQLColumnMappingOverrides
-
-class MergeIntoNotMatchedBySourceCDCPart1SQLPathBasedDVsPredPushOffSuite
-  extends MergeIntoNotMatchedBySourceCDCPart1Tests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with MergeIntoDVsMixin
-  with PredicatePushdownDisabled
-
-class MergeIntoNotMatchedBySourceCDCPart1SQLPathBasedDVsPredPushOnSuite
-  extends MergeIntoNotMatchedBySourceCDCPart1Tests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with MergeIntoDVsMixin
-  with PredicatePushdownEnabled
-
-class MergeIntoNotMatchedBySourceCDCPart1SQLPathBasedCDCOnSuite
-  extends MergeIntoNotMatchedBySourceCDCPart1Tests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeCDCMixin
-
-class MergeIntoNotMatchedBySourceCDCPart1SQLPathBasedCDCOnDVsPredPushOffSuite
-  extends MergeIntoNotMatchedBySourceCDCPart1Tests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeIntoDVsMixin
-  with PredicatePushdownDisabled
-  with MergeCDCMixin
-  with MergeCDCWithDVsMixin
-
-class MergeIntoNotMatchedBySourceCDCPart1SQLPathBasedCDCOnDVsPredPushOnSuite
-  extends MergeIntoNotMatchedBySourceCDCPart1Tests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeIntoDVsMixin
-  with PredicatePushdownEnabled
-  with MergeCDCMixin
-  with MergeCDCWithDVsMixin
-
-class MergeIntoNotMatchedBySourceCDCPart2SQLNameBasedSuite
-  extends MergeIntoNotMatchedBySourceCDCPart2Tests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsNameBased
-
-class MergeIntoNotMatchedBySourceCDCPart2SQLPathBasedSuite
-  extends MergeIntoNotMatchedBySourceCDCPart2Tests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-
-class MergeIntoNotMatchedBySourceCDCPart2SQLPathBasedColMapIdModeSuite
-  extends MergeIntoNotMatchedBySourceCDCPart2Tests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with DeltaColumnMappingEnableIdMode
-  with MergeIntoSQLColumnMappingOverrides
-
-class MergeIntoNotMatchedBySourceCDCPart2SQLPathBasedColMapNameModeSuite
-  extends MergeIntoNotMatchedBySourceCDCPart2Tests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with DeltaColumnMappingEnableNameMode
-  with MergeIntoSQLColumnMappingOverrides
-
-class MergeIntoNotMatchedBySourceCDCPart2SQLPathBasedDVsPredPushOffSuite
-  extends MergeIntoNotMatchedBySourceCDCPart2Tests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with MergeIntoDVsMixin
-  with PredicatePushdownDisabled
-
-class MergeIntoNotMatchedBySourceCDCPart2SQLPathBasedDVsPredPushOnSuite
-  extends MergeIntoNotMatchedBySourceCDCPart2Tests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with MergeIntoDVsMixin
-  with PredicatePushdownEnabled
-
-class MergeIntoNotMatchedBySourceCDCPart2SQLPathBasedCDCOnSuite
-  extends MergeIntoNotMatchedBySourceCDCPart2Tests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeCDCMixin
-
-class MergeIntoNotMatchedBySourceCDCPart2SQLPathBasedCDCOnDVsPredPushOffSuite
-  extends MergeIntoNotMatchedBySourceCDCPart2Tests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeIntoDVsMixin
-  with PredicatePushdownDisabled
-  with MergeCDCMixin
-  with MergeCDCWithDVsMixin
-
-class MergeIntoNotMatchedBySourceCDCPart2SQLPathBasedCDCOnDVsPredPushOnSuite
-  extends MergeIntoNotMatchedBySourceCDCPart2Tests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeIntoDVsMixin
-  with PredicatePushdownEnabled
-  with MergeCDCMixin
-  with MergeCDCWithDVsMixin
-
-class MergeIntoSchemaEvolutionCoreSQLNameBasedSuite
-  extends MergeIntoSchemaEvolutionCoreTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsNameBased
-
-class MergeIntoSchemaEvolutionCoreSQLPathBasedSuite
-  extends MergeIntoSchemaEvolutionCoreTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-
-class MergeIntoSchemaEvolutionCoreSQLPathBasedColMapIdModeSuite
-  extends MergeIntoSchemaEvolutionCoreTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with DeltaColumnMappingEnableIdMode
-  with MergeIntoSQLColumnMappingOverrides
-
-class MergeIntoSchemaEvolutionCoreSQLPathBasedColMapNameModeSuite
-  extends MergeIntoSchemaEvolutionCoreTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with DeltaColumnMappingEnableNameMode
-  with MergeIntoSQLColumnMappingOverrides
-
-class MergeIntoSchemaEvolutionCoreSQLPathBasedDVsPredPushOffSuite
-  extends MergeIntoSchemaEvolutionCoreTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with MergeIntoDVsMixin
-  with PredicatePushdownDisabled
-
-class MergeIntoSchemaEvolutionCoreSQLPathBasedDVsPredPushOnSuite
-  extends MergeIntoSchemaEvolutionCoreTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with MergeIntoDVsMixin
-  with PredicatePushdownEnabled
-
-class MergeIntoSchemaEvolutionCoreSQLPathBasedCDCOnSuite
-  extends MergeIntoSchemaEvolutionCoreTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeCDCMixin
-
-class MergeIntoSchemaEvolutionCoreSQLPathBasedCDCOnDVsPredPushOffSuite
-  extends MergeIntoSchemaEvolutionCoreTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeIntoDVsMixin
-  with PredicatePushdownDisabled
-  with MergeCDCMixin
-  with MergeCDCWithDVsMixin
-
-class MergeIntoSchemaEvolutionCoreSQLPathBasedCDCOnDVsPredPushOnSuite
-  extends MergeIntoSchemaEvolutionCoreTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeIntoDVsMixin
-  with PredicatePushdownEnabled
-  with MergeCDCMixin
-  with MergeCDCWithDVsMixin
-
-class MergeIntoSchemaEvolutionBaseSQLNameBasedSuite
-  extends MergeIntoSchemaEvolutionBaseTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsNameBased
-
-class MergeIntoSchemaEvolutionBaseSQLPathBasedSuite
-  extends MergeIntoSchemaEvolutionBaseTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-
-class MergeIntoSchemaEvolutionBaseSQLPathBasedColMapIdModeSuite
-  extends MergeIntoSchemaEvolutionBaseTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with DeltaColumnMappingEnableIdMode
-  with MergeIntoSQLColumnMappingOverrides
-
-class MergeIntoSchemaEvolutionBaseSQLPathBasedColMapNameModeSuite
-  extends MergeIntoSchemaEvolutionBaseTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with DeltaColumnMappingEnableNameMode
-  with MergeIntoSQLColumnMappingOverrides
-
-class MergeIntoSchemaEvolutionBaseSQLPathBasedDVsPredPushOffSuite
-  extends MergeIntoSchemaEvolutionBaseTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with MergeIntoDVsMixin
-  with PredicatePushdownDisabled
-
-class MergeIntoSchemaEvolutionBaseSQLPathBasedDVsPredPushOnSuite
-  extends MergeIntoSchemaEvolutionBaseTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with MergeIntoDVsMixin
-  with PredicatePushdownEnabled
-
-class MergeIntoSchemaEvolutionBaseSQLPathBasedCDCOnSuite
-  extends MergeIntoSchemaEvolutionBaseTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeCDCMixin
-
-class MergeIntoSchemaEvolutionBaseSQLPathBasedCDCOnDVsPredPushOffSuite
-  extends MergeIntoSchemaEvolutionBaseTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeIntoDVsMixin
-  with PredicatePushdownDisabled
-  with MergeCDCMixin
-  with MergeCDCWithDVsMixin
-
-class MergeIntoSchemaEvolutionBaseSQLPathBasedCDCOnDVsPredPushOnSuite
-  extends MergeIntoSchemaEvolutionBaseTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeIntoDVsMixin
-  with PredicatePushdownEnabled
-  with MergeCDCMixin
-  with MergeCDCWithDVsMixin
-
-class MergeIntoSchemaEvolutionStoreAssignmentPolicySQLNameBasedSuite
-  extends MergeIntoSchemaEvolutionStoreAssignmentPolicyTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsNameBased
-
-class MergeIntoSchemaEvolutionStoreAssignmentPolicySQLPathBasedSuite
-  extends MergeIntoSchemaEvolutionStoreAssignmentPolicyTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-
-class MergeIntoSchemaEvolutionStoreAssignmentPolicySQLPathBasedColMapIdModeSuite
-  extends MergeIntoSchemaEvolutionStoreAssignmentPolicyTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with DeltaColumnMappingEnableIdMode
-  with MergeIntoSQLColumnMappingOverrides
-
-class MergeIntoSchemaEvolutionStoreAssignmentPolicySQLPathBasedColMapNameModeSuite
-  extends MergeIntoSchemaEvolutionStoreAssignmentPolicyTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with DeltaColumnMappingEnableNameMode
-  with MergeIntoSQLColumnMappingOverrides
-
-class MergeIntoSchemaEvolutionStoreAssignmentPolicySQLPathBasedDVsPredPushOffSuite
-  extends MergeIntoSchemaEvolutionStoreAssignmentPolicyTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with MergeIntoDVsMixin
-  with PredicatePushdownDisabled
-
-class MergeIntoSchemaEvolutionStoreAssignmentPolicySQLPathBasedDVsPredPushOnSuite
-  extends MergeIntoSchemaEvolutionStoreAssignmentPolicyTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with MergeIntoDVsMixin
-  with PredicatePushdownEnabled
-
-class MergeIntoSchemaEvolutionStoreAssignmentPolicySQLPathBasedCDCOnSuite
-  extends MergeIntoSchemaEvolutionStoreAssignmentPolicyTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeCDCMixin
-
-class MergeIntoSchemaEvolutionStoreAssignmentPolicySQLPathBasedCDCOnDVsPredPushOffSuite
-  extends MergeIntoSchemaEvolutionStoreAssignmentPolicyTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeIntoDVsMixin
-  with PredicatePushdownDisabled
-  with MergeCDCMixin
-  with MergeCDCWithDVsMixin
-
-class MergeIntoSchemaEvolutionStoreAssignmentPolicySQLPathBasedCDCOnDVsPredPushOnSuite
-  extends MergeIntoSchemaEvolutionStoreAssignmentPolicyTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeIntoDVsMixin
-  with PredicatePushdownEnabled
-  with MergeCDCMixin
-  with MergeCDCWithDVsMixin
-
-class MergeIntoSchemaEvolutionNotMatchedBySourceSQLNameBasedSuite
-  extends MergeIntoSchemaEvolutionNotMatchedBySourceTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsNameBased
-
-class MergeIntoSchemaEvolutionNotMatchedBySourceSQLPathBasedSuite
-  extends MergeIntoSchemaEvolutionNotMatchedBySourceTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-
-class MergeIntoSchemaEvolutionNotMatchedBySourceSQLPathBasedColMapIdModeSuite
-  extends MergeIntoSchemaEvolutionNotMatchedBySourceTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with DeltaColumnMappingEnableIdMode
-  with MergeIntoSQLColumnMappingOverrides
-
-class MergeIntoSchemaEvolutionNotMatchedBySourceSQLPathBasedColMapNameModeSuite
-  extends MergeIntoSchemaEvolutionNotMatchedBySourceTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with DeltaColumnMappingEnableNameMode
-  with MergeIntoSQLColumnMappingOverrides
-
-class MergeIntoSchemaEvolutionNotMatchedBySourceSQLPathBasedDVsPredPushOffSuite
-  extends MergeIntoSchemaEvolutionNotMatchedBySourceTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with MergeIntoDVsMixin
-  with PredicatePushdownDisabled
-
-class MergeIntoSchemaEvolutionNotMatchedBySourceSQLPathBasedDVsPredPushOnSuite
-  extends MergeIntoSchemaEvolutionNotMatchedBySourceTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with MergeIntoDVsMixin
-  with PredicatePushdownEnabled
-
-class MergeIntoSchemaEvolutionNotMatchedBySourceSQLPathBasedCDCOnSuite
-  extends MergeIntoSchemaEvolutionNotMatchedBySourceTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeCDCMixin
-
-class MergeIntoSchemaEvolutionNotMatchedBySourceSQLPathBasedCDCOnDVsPredPushOffSuite
-  extends MergeIntoSchemaEvolutionNotMatchedBySourceTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeIntoDVsMixin
-  with PredicatePushdownDisabled
-  with MergeCDCMixin
-  with MergeCDCWithDVsMixin
-
-class MergeIntoSchemaEvolutionNotMatchedBySourceSQLPathBasedCDCOnDVsPredPushOnSuite
-  extends MergeIntoSchemaEvolutionNotMatchedBySourceTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeIntoDVsMixin
-  with PredicatePushdownEnabled
-  with MergeCDCMixin
-  with MergeCDCWithDVsMixin
-
-class MergeIntoNestedStructInMapEvolutionSQLNameBasedSuite
-  extends MergeIntoNestedStructInMapEvolutionTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsNameBased
-
-class MergeIntoNestedStructInMapEvolutionSQLPathBasedSuite
-  extends MergeIntoNestedStructInMapEvolutionTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-
-class MergeIntoNestedStructInMapEvolutionSQLPathBasedColMapIdModeSuite
-  extends MergeIntoNestedStructInMapEvolutionTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with DeltaColumnMappingEnableIdMode
-  with MergeIntoSQLColumnMappingOverrides
-
-class MergeIntoNestedStructInMapEvolutionSQLPathBasedColMapNameModeSuite
-  extends MergeIntoNestedStructInMapEvolutionTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with DeltaColumnMappingEnableNameMode
-  with MergeIntoSQLColumnMappingOverrides
-
-class MergeIntoNestedStructInMapEvolutionSQLPathBasedDVsPredPushOffSuite
-  extends MergeIntoNestedStructInMapEvolutionTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with MergeIntoDVsMixin
-  with PredicatePushdownDisabled
-
-class MergeIntoNestedStructInMapEvolutionSQLPathBasedDVsPredPushOnSuite
-  extends MergeIntoNestedStructInMapEvolutionTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with MergeIntoDVsMixin
-  with PredicatePushdownEnabled
-
-class MergeIntoNestedStructInMapEvolutionSQLPathBasedCDCOnSuite
-  extends MergeIntoNestedStructInMapEvolutionTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeCDCMixin
-
-class MergeIntoNestedStructInMapEvolutionSQLPathBasedCDCOnDVsPredPushOffSuite
-  extends MergeIntoNestedStructInMapEvolutionTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeIntoDVsMixin
-  with PredicatePushdownDisabled
-  with MergeCDCMixin
-  with MergeCDCWithDVsMixin
-
-class MergeIntoNestedStructInMapEvolutionSQLPathBasedCDCOnDVsPredPushOnSuite
-  extends MergeIntoNestedStructInMapEvolutionTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeIntoDVsMixin
-  with PredicatePushdownEnabled
-  with MergeCDCMixin
-  with MergeCDCWithDVsMixin
-
-class MergeIntoNestedStructEvolutionSQLNameBasedSuite
-  extends MergeIntoNestedStructEvolutionTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsNameBased
-
-class MergeIntoNestedStructEvolutionSQLPathBasedSuite
-  extends MergeIntoNestedStructEvolutionTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-
-class MergeIntoNestedStructEvolutionSQLPathBasedColMapIdModeSuite
-  extends MergeIntoNestedStructEvolutionTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with DeltaColumnMappingEnableIdMode
-  with MergeIntoSQLColumnMappingOverrides
-
-class MergeIntoNestedStructEvolutionSQLPathBasedColMapNameModeSuite
-  extends MergeIntoNestedStructEvolutionTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with DeltaColumnMappingEnableNameMode
-  with MergeIntoSQLColumnMappingOverrides
-
-class MergeIntoNestedStructEvolutionSQLPathBasedDVsPredPushOffSuite
-  extends MergeIntoNestedStructEvolutionTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with MergeIntoDVsMixin
-  with PredicatePushdownDisabled
-
-class MergeIntoNestedStructEvolutionSQLPathBasedDVsPredPushOnSuite
-  extends MergeIntoNestedStructEvolutionTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with MergeIntoDVsMixin
-  with PredicatePushdownEnabled
-
-class MergeIntoNestedStructEvolutionSQLPathBasedCDCOnSuite
-  extends MergeIntoNestedStructEvolutionTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeCDCMixin
-
-class MergeIntoNestedStructEvolutionSQLPathBasedCDCOnDVsPredPushOffSuite
-  extends MergeIntoNestedStructEvolutionTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeIntoDVsMixin
-  with PredicatePushdownDisabled
-  with MergeCDCMixin
-  with MergeCDCWithDVsMixin
-
-class MergeIntoNestedStructEvolutionSQLPathBasedCDCOnDVsPredPushOnSuite
-  extends MergeIntoNestedStructEvolutionTests
-  with MergeIntoSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with MergeIntoDVsMixin
-  with PredicatePushdownEnabled
-  with MergeCDCMixin
-  with MergeCDCWithDVsMixin
-
-class MergeIntoMaterializeSourceMergePersistentDVOffSuite
-  extends MergeIntoMaterializeSourceTests
-  with MergePersistentDVDisabled
-
-class MergeIntoMaterializeSourceErrorMergePersistentDVOffSuite
-  extends MergeIntoMaterializeSourceErrorTests
-  with MergePersistentDVDisabled
-
-class RowTrackingMergeCommonNameBasedSuite
-  extends RowTrackingMergeCommonTests
-  with DeltaDMLTestUtilsNameBased
-
-class RowTrackingMergeCommonNameBasedCDCOnSuite
-  extends RowTrackingMergeCommonTests
-  with DeltaDMLTestUtilsNameBased
-  with CDCEnabled
-
-class RowTrackingMergeCommonNameBasedRowTrackingMergeDVSuite
-  extends RowTrackingMergeCommonTests
-  with DeltaDMLTestUtilsNameBased
-  with RowTrackingMergeDVMixin
-
-class RowTrackingMergeCommonNameBasedCDCOnRowTrackingMergeDVSuite
-  extends RowTrackingMergeCommonTests
-  with DeltaDMLTestUtilsNameBased
-  with CDCEnabled
-  with RowTrackingMergeDVMixin
-
-class RowTrackingMergeCommonNameBasedDVOffMergePersistentDVOffSuite
-  extends RowTrackingMergeCommonTests
-  with DeltaDMLTestUtilsNameBased
-  with PersistentDVDisabled
-  with MergePersistentDVDisabled
-
-class RowTrackingMergeCommonNameBasedCDCOnDVOffMergePersistentDVOffSuite
-  extends RowTrackingMergeCommonTests
-  with DeltaDMLTestUtilsNameBased
-  with CDCEnabled
-  with PersistentDVDisabled
-  with MergePersistentDVDisabled
-
-class RowTrackingMergeCommonNameBasedColMapIdModeSuite
-  extends RowTrackingMergeCommonTests
-  with DeltaDMLTestUtilsNameBased
-  with DeltaColumnMappingEnableIdMode
-
-class RowTrackingMergeCommonNameBasedColMapNameModeSuite
-  extends RowTrackingMergeCommonTests
-  with DeltaDMLTestUtilsNameBased
-  with DeltaColumnMappingEnableNameMode
-
-class RowTrackingMergeCommonNameBasedColMapIdModeCDCOnRowTrackingMergeDVSuite
-  extends RowTrackingMergeCommonTests
-  with DeltaDMLTestUtilsNameBased
-  with DeltaColumnMappingEnableIdMode
-  with CDCEnabled
-  with RowTrackingMergeDVMixin
-
-class RowTrackingMergeCommonNameBasedColMapNameModeCDCOnRowTrackingMergeDVSuite
-  extends RowTrackingMergeCommonTests
-  with DeltaDMLTestUtilsNameBased
-  with DeltaColumnMappingEnableNameMode
-  with CDCEnabled
-  with RowTrackingMergeDVMixin
