+++ spark/src/test/scala/org/apache/spark/sql/delta/generatedsuites/MergeSuitesMergeIntoSQLTests.scala
@@ -1,0 +1,98 @@
+/*
+ * Copyright (2021) The Delta Lake Project Authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+// ***********************************************************************************
+// * This file is automatically generated. Manual modification is not allowed.       *
+// * There is a unit test that should prevent merging a manual change.               *
+// *                                                                                 *
+// * To make changes to the suites, modify the generator script config at            *
+// * SuiteGeneratorConfig.scala and run it. The generator can be run via the         *
+// * sbt command deltaSuiteGenerator / run.                                          *
+// *                                                                                 *
+// * DO NOT TOUCH ANYTHING IN THIS FILE!                                             *
+// ***********************************************************************************
+
+// scalastyle:off line.size.limit
+package org.apache.spark.sql.delta.generatedsuites
+
+import org.apache.spark.sql.delta._
+import org.apache.spark.sql.delta.cdc._
+import org.apache.spark.sql.delta.rowid._
+
+class MergeIntoSQLSQLNameBasedSuite
+  extends MergeIntoSQLTests
+  with MergeIntoSQLMixin
+  with DeltaDMLTestUtilsNameBased
+
+class MergeIntoSQLSQLPathBasedCDCOnDVsPredPushOffSuite
+  extends MergeIntoSQLTests
+  with MergeIntoSQLMixin
+  with DeltaDMLTestUtilsPathBased
+  with CDCEnabled
+  with MergeIntoDVsMixin
+  with PredicatePushdownDisabled
+  with MergeCDCMixin
+  with MergeCDCWithDVsMixin
+
+class MergeIntoSQLSQLPathBasedCDCOnDVsPredPushOnSuite
+  extends MergeIntoSQLTests
+  with MergeIntoSQLMixin
+  with DeltaDMLTestUtilsPathBased
+  with CDCEnabled
+  with MergeIntoDVsMixin
+  with PredicatePushdownEnabled
+  with MergeCDCMixin
+  with MergeCDCWithDVsMixin
+
+class MergeIntoSQLSQLPathBasedCDCOnSuite
+  extends MergeIntoSQLTests
+  with MergeIntoSQLMixin
+  with DeltaDMLTestUtilsPathBased
+  with CDCEnabled
+  with MergeCDCMixin
+
+class MergeIntoSQLSQLPathBasedColMapIdModeSuite
+  extends MergeIntoSQLTests
+  with MergeIntoSQLMixin
+  with DeltaDMLTestUtilsPathBased
+  with DeltaColumnMappingEnableIdMode
+  with MergeIntoSQLColumnMappingOverrides
+
+class MergeIntoSQLSQLPathBasedColMapNameModeSuite
+  extends MergeIntoSQLTests
+  with MergeIntoSQLMixin
+  with DeltaDMLTestUtilsPathBased
+  with DeltaColumnMappingEnableNameMode
+  with MergeIntoSQLColumnMappingOverrides
+
+class MergeIntoSQLSQLPathBasedDVsPredPushOffSuite
+  extends MergeIntoSQLTests
+  with MergeIntoSQLMixin
+  with DeltaDMLTestUtilsPathBased
+  with MergeIntoDVsMixin
+  with PredicatePushdownDisabled
+
+class MergeIntoSQLSQLPathBasedDVsPredPushOnSuite
+  extends MergeIntoSQLTests
+  with MergeIntoSQLMixin
+  with DeltaDMLTestUtilsPathBased
+  with MergeIntoDVsMixin
+  with PredicatePushdownEnabled
+
+class MergeIntoSQLSQLPathBasedSuite
+  extends MergeIntoSQLTests
+  with MergeIntoSQLMixin
+  with DeltaDMLTestUtilsPathBased
