+++ spark/src/test/scala/org/apache/spark/sql/delta/generatedsuites/MergeSuitesRowTrackingMergeCommonTests.scala
@@ -1,0 +1,90 @@
+/*
+ * Copyright (2021) The Delta Lake Project Authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+// ***********************************************************************************
+// * This file is automatically generated. Manual modification is not allowed.       *
+// * There is a unit test that should prevent merging a manual change.               *
+// *                                                                                 *
+// * To make changes to the suites, modify the generator script config at            *
+// * SuiteGeneratorConfig.scala and run it. The generator can be run via the         *
+// * sbt command deltaSuiteGenerator / run.                                          *
+// *                                                                                 *
+// * DO NOT TOUCH ANYTHING IN THIS FILE!                                             *
+// ***********************************************************************************
+
+// scalastyle:off line.size.limit
+package org.apache.spark.sql.delta.generatedsuites
+
+import org.apache.spark.sql.delta._
+import org.apache.spark.sql.delta.cdc._
+import org.apache.spark.sql.delta.rowid._
+
+class RowTrackingMergeCommonNameBasedCDCOnDVOffMergePersistentDVOffSuite
+  extends RowTrackingMergeCommonTests
+  with DeltaDMLTestUtilsNameBased
+  with CDCEnabled
+  with PersistentDVDisabled
+  with MergePersistentDVDisabled
+
+class RowTrackingMergeCommonNameBasedCDCOnRowTrackingMergeDVSuite
+  extends RowTrackingMergeCommonTests
+  with DeltaDMLTestUtilsNameBased
+  with CDCEnabled
+  with RowTrackingMergeDVMixin
+
+class RowTrackingMergeCommonNameBasedCDCOnSuite
+  extends RowTrackingMergeCommonTests
+  with DeltaDMLTestUtilsNameBased
+  with CDCEnabled
+
+class RowTrackingMergeCommonNameBasedColMapIdModeCDCOnRowTrackingMergeDVSuite
+  extends RowTrackingMergeCommonTests
+  with DeltaDMLTestUtilsNameBased
+  with DeltaColumnMappingEnableIdMode
+  with CDCEnabled
+  with RowTrackingMergeDVMixin
+
+class RowTrackingMergeCommonNameBasedColMapIdModeSuite
+  extends RowTrackingMergeCommonTests
+  with DeltaDMLTestUtilsNameBased
+  with DeltaColumnMappingEnableIdMode
+
+class RowTrackingMergeCommonNameBasedColMapNameModeCDCOnRowTrackingMergeDVSuite
+  extends RowTrackingMergeCommonTests
+  with DeltaDMLTestUtilsNameBased
+  with DeltaColumnMappingEnableNameMode
+  with CDCEnabled
+  with RowTrackingMergeDVMixin
+
+class RowTrackingMergeCommonNameBasedColMapNameModeSuite
+  extends RowTrackingMergeCommonTests
+  with DeltaDMLTestUtilsNameBased
+  with DeltaColumnMappingEnableNameMode
+
+class RowTrackingMergeCommonNameBasedDVOffMergePersistentDVOffSuite
+  extends RowTrackingMergeCommonTests
+  with DeltaDMLTestUtilsNameBased
+  with PersistentDVDisabled
+  with MergePersistentDVDisabled
+
+class RowTrackingMergeCommonNameBasedRowTrackingMergeDVSuite
+  extends RowTrackingMergeCommonTests
+  with DeltaDMLTestUtilsNameBased
+  with RowTrackingMergeDVMixin
+
+class RowTrackingMergeCommonNameBasedSuite
+  extends RowTrackingMergeCommonTests
+  with DeltaDMLTestUtilsNameBased
