--- spark/src/test/scala/org/apache/spark/sql/delta/generatedsuites/UpdateSuites1.scala
@@ -1,330 +1,0 @@
-/*
- * Copyright (2021) The Delta Lake Project Authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-// ***********************************************************************************
-// * This file is automatically generated. Manual modification is not allowed.       *
-// * There is a unit test that should prevent merging a manual change.               *
-// *                                                                                 *
-// * To make changes to the suites, modify the generator script config at            *
-// * SuiteGeneratorConfig.scala and run it. The generator can be run via the         *
-// * sbt command deltaSuiteGenerator / run.                                          *
-// *                                                                                 *
-// * DO NOT TOUCH ANYTHING IN THIS FILE!                                             *
-// ***********************************************************************************
-
-// scalastyle:off line.size.limit
-package org.apache.spark.sql.delta.generatedsuites
-
-import org.apache.spark.sql.delta._
-import org.apache.spark.sql.delta.cdc._
-import org.apache.spark.sql.delta.rowid._
-import org.apache.spark.sql.delta.rowtracking._
-
-class UpdateScalaScalaSuite extends UpdateScalaTests with UpdateScalaMixin
-class UpdateBaseMiscScalaSuite extends UpdateBaseMiscTests with UpdateScalaMixin
-
-class UpdateSQLSQLNameBasedSuite
-  extends UpdateSQLTests
-  with UpdateSQLMixin
-  with DeltaDMLTestUtilsNameBased
-
-class UpdateBaseTempViewSQLNameBasedSuite
-  extends UpdateBaseTempViewTests
-  with UpdateSQLMixin
-  with DeltaDMLTestUtilsNameBased
-
-class UpdateBaseMiscSQLNameBasedSuite
-  extends UpdateBaseMiscTests
-  with UpdateSQLMixin
-  with DeltaDMLTestUtilsNameBased
-
-class UpdateCDCWithDeletionVectorsSQLPathBasedCDCOnDVSuite
-  extends UpdateCDCWithDeletionVectorsTests
-  with UpdateSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with UpdateSQLWithDeletionVectorsMixin
-
-class UpdateCDCSQLPathBasedCDCOnSuite
-  extends UpdateCDCTests
-  with UpdateSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-
-class UpdateCDCSQLPathBasedCDCOnRowTrackingOffSuite
-  extends UpdateCDCTests
-  with UpdateSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with RowTrackingDisabled
-
-class UpdateCDCSQLPathBasedCDCOnRowTrackingOnSuite
-  extends UpdateCDCTests
-  with UpdateSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with RowTrackingEnabled
-  with UpdateWithRowTrackingOverrides
-
-class UpdateCDCSQLPathBasedCDCOnDVSuite
-  extends UpdateCDCTests
-  with UpdateSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with UpdateSQLWithDeletionVectorsMixin
-
-class UpdateSQLWithDeletionVectorsSQLPathBasedCDCOnDVSuite
-  extends UpdateSQLWithDeletionVectorsTests
-  with UpdateSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with UpdateSQLWithDeletionVectorsMixin
-
-class UpdateSQLWithDeletionVectorsSQLPathBasedDVPredPushOffSuite
-  extends UpdateSQLWithDeletionVectorsTests
-  with UpdateSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with UpdateSQLWithDeletionVectorsMixin
-  with PredicatePushdownDisabled
-
-class UpdateSQLWithDeletionVectorsSQLPathBasedDVPredPushOnSuite
-  extends UpdateSQLWithDeletionVectorsTests
-  with UpdateSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with UpdateSQLWithDeletionVectorsMixin
-  with PredicatePushdownEnabled
-
-class UpdateSQLSQLPathBasedSuite
-  extends UpdateSQLTests
-  with UpdateSQLMixin
-  with DeltaDMLTestUtilsPathBased
-
-class UpdateSQLSQLPathBasedCDCOnSuite
-  extends UpdateSQLTests
-  with UpdateSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-
-class UpdateSQLSQLPathBasedRowTrackingOffSuite
-  extends UpdateSQLTests
-  with UpdateSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with RowTrackingDisabled
-
-class UpdateSQLSQLPathBasedRowTrackingOnSuite
-  extends UpdateSQLTests
-  with UpdateSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with RowTrackingEnabled
-  with UpdateWithRowTrackingOverrides
-
-class UpdateSQLSQLPathBasedCDCOnRowTrackingOffSuite
-  extends UpdateSQLTests
-  with UpdateSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with RowTrackingDisabled
-
-class UpdateSQLSQLPathBasedCDCOnRowTrackingOnSuite
-  extends UpdateSQLTests
-  with UpdateSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with RowTrackingEnabled
-  with UpdateWithRowTrackingOverrides
-
-class UpdateSQLSQLPathBasedCDCOnDVSuite
-  extends UpdateSQLTests
-  with UpdateSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with UpdateSQLWithDeletionVectorsMixin
-
-class UpdateSQLSQLPathBasedDVPredPushOffSuite
-  extends UpdateSQLTests
-  with UpdateSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with UpdateSQLWithDeletionVectorsMixin
-  with PredicatePushdownDisabled
-
-class UpdateSQLSQLPathBasedDVPredPushOnSuite
-  extends UpdateSQLTests
-  with UpdateSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with UpdateSQLWithDeletionVectorsMixin
-  with PredicatePushdownEnabled
-
-class UpdateBaseTempViewSQLPathBasedSuite
-  extends UpdateBaseTempViewTests
-  with UpdateSQLMixin
-  with DeltaDMLTestUtilsPathBased
-
-class UpdateBaseTempViewSQLPathBasedCDCOnSuite
-  extends UpdateBaseTempViewTests
-  with UpdateSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-
-class UpdateBaseTempViewSQLPathBasedRowTrackingOffSuite
-  extends UpdateBaseTempViewTests
-  with UpdateSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with RowTrackingDisabled
-
-class UpdateBaseTempViewSQLPathBasedRowTrackingOnSuite
-  extends UpdateBaseTempViewTests
-  with UpdateSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with RowTrackingEnabled
-  with UpdateWithRowTrackingOverrides
-
-class UpdateBaseTempViewSQLPathBasedCDCOnRowTrackingOffSuite
-  extends UpdateBaseTempViewTests
-  with UpdateSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with RowTrackingDisabled
-
-class UpdateBaseTempViewSQLPathBasedCDCOnRowTrackingOnSuite
-  extends UpdateBaseTempViewTests
-  with UpdateSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with RowTrackingEnabled
-  with UpdateWithRowTrackingOverrides
-
-class UpdateBaseTempViewSQLPathBasedCDCOnDVSuite
-  extends UpdateBaseTempViewTests
-  with UpdateSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with UpdateSQLWithDeletionVectorsMixin
-
-class UpdateBaseTempViewSQLPathBasedDVPredPushOffSuite
-  extends UpdateBaseTempViewTests
-  with UpdateSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with UpdateSQLWithDeletionVectorsMixin
-  with PredicatePushdownDisabled
-
-class UpdateBaseTempViewSQLPathBasedDVPredPushOnSuite
-  extends UpdateBaseTempViewTests
-  with UpdateSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with UpdateSQLWithDeletionVectorsMixin
-  with PredicatePushdownEnabled
-
-class UpdateBaseMiscSQLPathBasedSuite
-  extends UpdateBaseMiscTests
-  with UpdateSQLMixin
-  with DeltaDMLTestUtilsPathBased
-
-class UpdateBaseMiscSQLPathBasedCDCOnSuite
-  extends UpdateBaseMiscTests
-  with UpdateSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-
-class UpdateBaseMiscSQLPathBasedRowTrackingOffSuite
-  extends UpdateBaseMiscTests
-  with UpdateSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with RowTrackingDisabled
-
-class UpdateBaseMiscSQLPathBasedRowTrackingOnSuite
-  extends UpdateBaseMiscTests
-  with UpdateSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with RowTrackingEnabled
-  with UpdateWithRowTrackingOverrides
-
-class UpdateBaseMiscSQLPathBasedCDCOnRowTrackingOffSuite
-  extends UpdateBaseMiscTests
-  with UpdateSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with RowTrackingDisabled
-
-class UpdateBaseMiscSQLPathBasedCDCOnRowTrackingOnSuite
-  extends UpdateBaseMiscTests
-  with UpdateSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with RowTrackingEnabled
-  with UpdateWithRowTrackingOverrides
-
-class UpdateBaseMiscSQLPathBasedCDCOnDVSuite
-  extends UpdateBaseMiscTests
-  with UpdateSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with CDCEnabled
-  with UpdateSQLWithDeletionVectorsMixin
-
-class UpdateBaseMiscSQLPathBasedDVPredPushOffSuite
-  extends UpdateBaseMiscTests
-  with UpdateSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with UpdateSQLWithDeletionVectorsMixin
-  with PredicatePushdownDisabled
-
-class UpdateBaseMiscSQLPathBasedDVPredPushOnSuite
-  extends UpdateBaseMiscTests
-  with UpdateSQLMixin
-  with DeltaDMLTestUtilsPathBased
-  with UpdateSQLWithDeletionVectorsMixin
-  with PredicatePushdownEnabled
-
-class RowTrackingUpdateCommonSuite extends RowTrackingUpdateCommonTests
-class RowTrackingUpdateCommonCDCOnSuite extends RowTrackingUpdateCommonTests with CDCEnabled
-
-class RowTrackingUpdateCommonColMapIdModeSuite
-  extends RowTrackingUpdateCommonTests
-  with DeltaColumnMappingEnableIdMode
-
-class RowTrackingUpdateCommonColMapNameModeSuite
-  extends RowTrackingUpdateCommonTests
-  with DeltaColumnMappingEnableNameMode
-
-class RowTrackingUpdateCommonCDCOnColMapIdModeSuite
-  extends RowTrackingUpdateCommonTests
-  with CDCEnabled
-  with DeltaColumnMappingEnableIdMode
-
-class RowTrackingUpdateCommonCDCOnColMapNameModeSuite
-  extends RowTrackingUpdateCommonTests
-  with CDCEnabled
-  with DeltaColumnMappingEnableNameMode
-
-class RowTrackingUpdateCommonRowTrackingUpdateDVSuite
-  extends RowTrackingUpdateCommonTests
-  with RowTrackingUpdateDVMixin
-
-class RowTrackingUpdateCommonRowTrackingUpdateDVCDCOnSuite
-  extends RowTrackingUpdateCommonTests
-  with RowTrackingUpdateDVMixin
-  with CDCEnabled
-
-class RowTrackingUpdateCommonRowTrackingUpdateDVCDCOnColMapIdModeSuite
-  extends RowTrackingUpdateCommonTests
-  with RowTrackingUpdateDVMixin
-  with CDCEnabled
-  with DeltaColumnMappingEnableIdMode
-
-class RowTrackingUpdateCommonRowTrackingUpdateDVCDCOnColMapNameModeSuite
-  extends RowTrackingUpdateCommonTests
-  with RowTrackingUpdateDVMixin
-  with CDCEnabled
-  with DeltaColumnMappingEnableNameMode
